Extend project features add frontend and API documentation

+++ orders/api/views.py
@@ -4,9 +4,12 @@ from rest_framework.response import Response
 from rest_framework.permissions import IsAuthenticated
 from carts.models import Cart, Address
 from orders.models import Order, OrderItem
-from .serializers import OrderSerializer, CreateOrderSerializer
+from .serializers import OrderSerializer, CreateOrderSerializer, SellerOrderSerializer
 from rest_framework import generics, status
 from notifications.utils import create_notification
+# forseller
+from django.db.models import Prefetch
+from products.api import permissions
 
 class OrderListView(generics.ListAPIView):
     serializer_class = OrderSerializer
@@ -31,7 +34,8 @@ class CreateOrderView(generics.CreateAPIView):
         cart_items = cart.cart_products.select_related("product")
         if not cart_items.exists():
             return Response({"error": "Cart is empty"},status=status.HTTP_400_BAD_REQUEST)
-        order = Order.objects.create(user=request.user,address=address)
+        payment_method = serializer.validated_data["payment_method"]
+        order = Order.objects.create(user=request.user,address=address, payment_method=payment_method)
         total_price = 0
         for item in cart_items:
             product = item.product
@@ -51,6 +55,27 @@ class CreateOrderView(generics.CreateAPIView):
         order.save(update_fields=["total_price"])
         # this 1 line is for testing the notification system, you can remove it later
         create_notification(user=request.user,title="Order Confirmed", message=f"Your order #{order.id} has been placed successfully.", notif_type="order")
-        cart_items.delete()
+        if payment_method == "COD":
+            cart_items.delete()
         response_serializer = OrderSerializer(order)
         return Response(response_serializer.data,status=status.HTTP_201_CREATED)
+
+
+# for seller
+class OrderListView(generics.ListAPIView):
+    serializer_class = OrderSerializer
+    permission_classes = [IsAuthenticated]
+    def get_queryset(self):
+        return (Order.objects.filter(user=self.request.user).prefetch_related("items"))
+
+class SellerOrderListView(generics.ListAPIView):
+    serializer_class = SellerOrderSerializer
+    permission_classes = [IsAuthenticated, permissions.IsSellerOrReadOnly]
+    def get_queryset(self):
+        seller = self.request.user
+        seller_items = (OrderItem.objects.filter(product__seller=seller)
+            .select_related("product"))
+        return (Order.objects.filter(items__product__seller=seller).distinct()
+            .prefetch_related(Prefetch("items", queryset=seller_items))
+            .order_by("-created_at")
+        )
